+++ ydb/core/grpc_services/service_import.h
@@ -11,6 +11,7 @@ class IFacilityProvider;
 void DoImportFromS3Request(std::unique_ptr<IRequestOpCtx> p, const IFacilityProvider& f);
 void DoImportFromFsRequest(std::unique_ptr<IRequestOpCtx> p, const IFacilityProvider& f);
 void DoListObjectsInS3ExportRequest(std::unique_ptr<IRequestOpCtx> p, const IFacilityProvider& f);
+void DoListObjectsInFsExportRequest(std::unique_ptr<IRequestOpCtx> p, const IFacilityProvider& f);
 void DoImportDataRequest(std::unique_ptr<IRequestOpCtx> p, const IFacilityProvider& f);
 
 }