+++ ydb/core/wrappers/fs_storage_config.cpp
@@ -30,4 +30,9 @@ TFsExternalStorageConfig::TFsExternalStorageConfig(const Ydb::Import::ImportFrom
 {
 }
 
+TFsExternalStorageConfig::TFsExternalStorageConfig(const Ydb::Import::ListObjectsInFsExportSettings& settings)
+    : BasePath(settings.base_path())
+{
+}
+
 } // NKikimr::NWrappers::NExternalStorage