+++ ydb/public/lib/ydb_cli/commands/ydb_service_import.h
@@ -97,6 +97,7 @@ public:
     virtual int Run(TConfig& config) override;
 
     NImport::TImportFromFsSettings MakeImportSettings();
+    NImport::TListObjectsInFsExportSettings MakeListObjectsSettings();
 
 private:
     DEFINE_PARSEABLE_STRUCT(TItemNfs, TItemFields, Source, Destination);