+++ ydb/public/sdk/cpp/src/client/import/out.cpp
@@ -15,3 +15,11 @@ Y_DECLARE_OUT_SPEC(, NYdb::NImport::TListObjectsInS3ExportResult, o, x) {
 Y_DECLARE_OUT_SPEC(, NYdb::NImport::TListObjectsInS3ExportResult::TItem, o, x) {
     return x.Out(o);
 }
+
+Y_DECLARE_OUT_SPEC(, NYdb::NImport::TListObjectsInFsExportResult, o, x) {
+    return x.Out(o);
+}
+
+Y_DECLARE_OUT_SPEC(, NYdb::NImport::TListObjectsInFsExportResult::TItem, o, x) {
+    return x.Out(o);
+}